[EI-136] [api] refactor: src/util/type type명 변경
