+++ api/src/numerical-guidance/application/query/get-indicator-list/indicator-list.dto.ts
@@ -1,4 +1,4 @@
-import { Market } from 'src/type/market';
+import { Market } from 'src/utils/type/type';
 
 export class IndicatorListDto {
   indicatorList: Indicator[];
